Separate item tabs into unique files, rename local_app.py to main.py
